Update BLE_USB project path, adjust connection interval and characteristic UUID, comment out unused characteristics, optimize data parsing function

+++ APP/include/usbd_composite_km.h
@@ -38,6 +38,6 @@ extern void MCU_Sleep_Wakeup_Operate( void );
 extern void CH9329_SendAck(uint8_t addr, uint8_t cmd_code, uint8_t status);
 extern void ProcessSpecialKeys(uint8_t *keys, uint8_t keyCount);
 extern uint8_t MapSpecialKeyToModBit(uint8_t key);
-extern void CH9329_DataParser(uint8_t* buf, uint8_t len);
+extern void CH9329_DataParser(uint8_t* buf, uint16_t len);
 extern void USB_DataRx_To_KMHandle(void);
 #endif
